Implemented dotenv (see elsewhere for the env file), setup the MongoDB for flexsheet, added some missing packages, server now connects to mongoDB when run
There's also a script I've added to the package.json that will let you just add a user to the database for testing purposes.  It's easiest to run as "npm run mongoUser <username> <password>." This can be changed later down the line for additional testing, or more or less scrapped in favour of sending post requests to the backend.